Drop support for Debian Wheezy
Debian Wheezy is out of support by Debian for years now, so there's
little point in still supporting it here. Besides that,
`$lsbdistcodename` doesn't seem to work anymore on more current versions
of Puppet (i.e. the one in Debian Trixie), so this would need changing
anyway, which might not be easily possible to do in a way which supports
both the ancient version of Puppet from Debian Wheezy and the more
current one in Debian Trixie.